feat: add ability to override GITHUB_ACTOR
The default GITHUB_TOKEN used for performing git operations cannot
operate on a protected branch. To circumvent this limitation, one can
generate a Personal Access Token and use that to override the default.

In a shared build environment, this will limit triggering this workflow
to a single user which may not be ideal. Organizations can choose to
generate a PAT for an administrator and then override the GITHUB_ACTOR
variable, thereby allowing anyone to successfully trigger this workflow.